refactor: Use new version store
