+++ renderer/components/TokenForm/index.tsx
@@ -1,21 +1,21 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { State } from '../../state/reducer';
-import { getAccessToken } from '../../state/selectors';
-import { DispatchProps } from '../../redux/react';
-import { UpdateToken } from '../../actions';
+import State from '../../store/state';
+import { getAccessToken } from '../../store/selectors';
+import { Dispatch } from '../../store/types';
+import { updateToken } from '../../store/user-config/actions';
 
 export interface Props {
   token: string;
 }
-type AllProps = Props & DispatchProps;
+type AllProps = Props & { dispatch: Dispatch };
 
 const TokenForm = (props: AllProps) => {
   const { token: currentToken, dispatch } = props;
   let token = currentToken;
   const handleSubmit = (event: React.FormEvent<any>) => {
     event.preventDefault();
-    dispatch(UpdateToken(token));
+    dispatch(updateToken(token));
   };
   return (
     <form onSubmit={handleSubmit}>
