+++ renderer/index.tsx
@@ -2,35 +2,36 @@ import * as React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import Root from './components/Root';
-import history from './history';
-import configureStore from './old-store';
-import { isConfigLoaded, getAccessToken } from './state/selectors';
-import { LoadUserConfig, Push } from './actions';
+import { History } from 'history';
+import createHashHistory from 'history/createHashHistory';
+import configureStore from './store';
+import { isConfigLoaded, getAccessToken } from './store/selectors';
+import { loadUserConfig } from './store/user-config/actions';
+import { push } from './store/router/actions';
 import * as paths from './const/paths';
 
 import './common.scss';
 
+const history = createHashHistory();
+const store = configureStore({ history });
+
 const renderView = () => {
   render(
-    <Provider store={store}>
+    <Provider store={store as any}>
       <Root history={history} />
     </Provider>,
     document.getElementById('root'),
   );
 };
 
-const store = configureStore();
-
-store.dispatch(LoadUserConfig());
-
-const unsubscribe = store.subscribe(() => {
+store.dispatch(loadUserConfig()).then(() => {
   const state = store.getState();
   if (!isConfigLoaded(state)) {
     return;
   }
 
   if (!getAccessToken(state) && history.location.pathname !== paths.tokenRegistration) {
-    store.dispatch(Push(paths.tokenRegistration));
+    store.dispatch(push(paths.tokenRegistration));
   }
 
   if (document.readyState === 'pending') {
@@ -39,6 +40,4 @@ const unsubscribe = store.subscribe(() => {
   else {
     renderView();
   }
-
-  unsubscribe();
 });
